Refactoring to make the emits work, starting work on the updating section image in the backend
