+++ etc/fixtures/initial-seed/analyses.json
@@ -137,6 +137,7 @@
         "content": [{
           "type": "images-dataset",
           "dataset": "Pedigree",
+          "field": "Pedigree",
           "value": []
         }]
       }
